__LINE__ macro seems to be always an int
git-svn-id: https://bibdesk.svn.sourceforge.net/svnroot/bibdesk/trunk/bibdesk@16079 0fceea05-a30d-0410-8a8b-80ef821fb0a1

+++ BibDocument_DataSource.m
@@ -437,7 +437,7 @@ static void addSubmenuForURLsToItem(NSArray *urls, NSMenuItem *anItem) {
                             NSError *xerror = nil;
                             // we can always write xattrs; this doesn't alter the original file's content in any way, but fails if you have a really long abstract/annote
                             if([[SKNExtendedAttributeManager sharedNoSplitManager] setExtendedAttributeNamed:BDSK_BUNDLE_IDENTIFIER @".bibtexstring" toValue:[[pub bibTeXString] dataUsingEncoding:NSUTF8StringEncoding] atPath:path options:0 error:&xerror] == NO)
-                                NSLog(@"%@ line %ld: adding xattrs failed with error %@", __FILENAMEASNSSTRING__, (long)__LINE__, xerror);
+                                NSLog(@"%@ line %d: adding xattrs failed with error %@", __FILENAMEASNSSTRING__, __LINE__, xerror);
                         }
                         row = [rowIndexes indexGreaterThanIndex:row];
                     }
@@ -493,7 +493,7 @@ static void addSubmenuForURLsToItem(NSArray *urls, NSMenuItem *anItem) {
                                 NSError *xerror = nil;
                                 // we can always write xattrs; this doesn't alter the original file's content in any way, but fails if you have a really long abstract/annote
                                 if([[SKNExtendedAttributeManager sharedNoSplitManager] setExtendedAttributeNamed:BDSK_BUNDLE_IDENTIFIER @".bibtexstring" toValue:[[pub bibTeXString] dataUsingEncoding:NSUTF8StringEncoding] atPath:path options:0 error:&xerror] == NO)
-                                    NSLog(@"%@ line %ld: adding xattrs failed with error %@", __FILENAMEASNSSTRING__, (long)__LINE__, xerror);
+                                    NSLog(@"%@ line %d: adding xattrs failed with error %@", __FILENAMEASNSSTRING__, __LINE__, xerror);
                             }
                         }
                         row = [rowIndexes indexGreaterThanIndex:row];
